+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- refactor(plugins): improve error handling and logging for plugin activation failures. (`929a37c`)
+
 ## 1.8.29 - 2026-06-28
 
 - Update .gitignore and clean up build-docs-nav script. (`b67c989`)